test: share embedded-struct scenarios across embed_struct tests
Extract three reused inline models into scenarios: user_with_address
(basic_embedded_struct, root_model_with_embedded_field), user_with_zip_address
(embedded_struct_fields_codegen and the two partial-update tests), and
company_office_address (the two nested-embed tests).
